+++ x/auth/ante.go
@@ -1,7 +1,7 @@
 package auth
 
 import (
-	"bytes"
+	"encoding/hex"
 	"fmt"
 	posCrypto "github.com/pokt-network/posmint/crypto"
 	sdk "github.com/pokt-network/posmint/types"
@@ -23,114 +23,110 @@ func NewAnteHandler(ak keeper.Keeper) sdk.AnteHandler {
 		if !ok {
 			return newCtx, sdk.ErrInternal("tx must be StdTx").Result(), true
 		}
-		//Default Fee
-		defaultFee := sdk.NewCoin(sdk.DefaultStakeDenom, sdk.NewInt(100000))
-		//Check Tx Amount and Denomination.
-		if !stdTx.Fee.AmountOf(sdk.DefaultStakeDenom).Equal(defaultFee.Amount) {
-			return newCtx, sdk.ErrInsufficientFee(
-				fmt.Sprintf(
-					"insufficient fees; got: %q required: %q", stdTx.Fee, defaultFee,
-				),
-			).Result(), true
-		}
-		params := ak.GetParams(ctx)
-		// Ensure that the provided fees meet a minimum threshold for the validator,
-		// if this is a CheckTx. This is only for local mempool purposes, and thus
-		// is only ran on check tx.
-		if ctx.IsCheckTx() && !simulate {
-			res := EnsureSufficientMempoolFees(ctx, stdTx.Fee)
-			if !res.IsOK() {
-				return newCtx, res, true
-			}
-		}
-		if res := ValidateSigCount(stdTx, params); !res.IsOK() {
-			return newCtx, res, true
-		}
+		// validate the transaction
 		if err := tx.ValidateBasic(); err != nil {
 			return newCtx, err.Result(), true
 		}
-		if res := ValidateMemo(stdTx, params); !res.IsOK() {
-			return newCtx, res, true
+		err := ValidateTransaction(ctx, ak, stdTx, ak.GetParams(ctx), tmNode, txBz, simulate)
+		if err != nil {
+			return newCtx, err.Result(), true
 		}
-		// stdSigs contains the sequence number, account number, and signatures.
-		// When simulating, this would just be a 0-length slice.
-		signerAddrs := stdTx.GetSigners()
-		signerAccs := make([]Account, len(signerAddrs))
-		// fetch first signer, who's going to pay the fees
-		signerAccs[0], res = GetSignerAcc(ctx, ak, signerAddrs[0])
-		if !res.IsOK() {
-			return newCtx, res, true
-		}
-		// deduct the fees
-		if !stdTx.Fee.IsZero() {
-			res = DeductFees(ak, ctx, signerAccs[0], stdTx.Fee)
-			if !res.IsOK() {
-				return newCtx, res, true
-			}
-
-			// reload the account as fees have been deducted
-			signerAccs[0] = ak.GetAccount(ctx, signerAccs[0].GetAddress())
-		}
-		// stdSigs contains signatures.
-		// When simulating, this would just be a 0-length slice.
-		stdSigs := stdTx.GetSignatures()
-		for i := 0; i < len(stdSigs); i++ {
-			// skip the fee payer, account is cached and fees were deducted already
-			if i != 0 {
-				signerAccs[i], res = GetSignerAcc(ctx, ak, signerAddrs[i])
-				if !res.IsOK() {
-					return newCtx, res, true
-				}
-			}
-			// check signature, return account with incremented nonce
-			signBytes := GetSignBytes(ctx.ChainID(), stdTx)
-			signerAccs[i], res = processSig(signerAccs[i], stdSigs[i], signBytes, simulate)
-			// todo when editing tendermint, pass txIndexer so no http
-			c := client.NewHTTP(tmNode.Config().RPC.ListenAddress, "/websocket")
-			_, err := c.Tx(tmTypes.Tx(txBz).Hash(), false)
-			if err == nil {
-				return newCtx,
-					sdk.ErrUnauthorized(fmt.Sprint("transaction with this hash already found, possible replay attack, please try a different entropy")).Result(),
-					true
-			}
-			if !res.IsOK() {
-				return newCtx, res, true
-			}
-			ak.SetAccount(ctx, signerAccs[i])
+		err = DeductFees(ak, ctx, stdTx)
+		if err != nil {
+			return newCtx, err.Result(), true
 		}
 		return ctx, sdk.Result{}, false // continue...
 	}
 }
 
-// GetSignerAcc returns an account for a given address that is expected to sign
-// a transaction.
-func GetSignerAcc(ctx sdk.Ctx, ak keeper.Keeper, addr sdk.Address) (Account, sdk.Result) {
-	if acc := ak.GetAccount(ctx, addr); acc != nil {
-		return acc, sdk.Result{}
+func ValidateTransaction(ctx sdk.Ctx, k Keeper, stdTx StdTx, params Params, tmNode *node.Node, txBz []byte, simulate bool) sdk.Error {
+	// validate the memo
+	if err := ValidateMemo(stdTx, params); err != nil {
+		return types.ErrInvalidMemo(ModuleName, err)
 	}
-	return nil, sdk.ErrUnknownAddress(fmt.Sprintf("account %s does not exist", addr)).Result()
-}
-
-// ValidateSigCount validates that the transaction has a valid cumulative total
-// amount of signatures.
-func ValidateSigCount(stdTx StdTx, params Params) sdk.Result {
-	stdSigs := stdTx.GetSignatures()
-
-	sigCount := 0
-	for i := 0; i < len(stdSigs); i++ {
-		sigCount += CountSubKeys(stdSigs[i].PublicKey)
-		if uint64(sigCount) > params.TxSigLimit {
-			return sdk.ErrTooManySignatures(
-				fmt.Sprintf("signatures: %d, limit: %d", sigCount, params.TxSigLimit),
-			).Result()
+	var pk posCrypto.PublicKey
+	// attempt to get the public key from the signature
+	if stdTx.Signature.PublicKey != nil && len(stdTx.Signature.PublicKey.RawBytes()) != 0 {
+		pk = stdTx.Signature.PublicKey
+	} else {
+		// public key in the signature not found so check world state
+		acc := k.GetAccount(ctx, stdTx.GetSigner())
+		if acc == nil {
+			return types.ErrAccountNotFound(ModuleName)
+		}
+		if pk = acc.GetPubKey(); pk == nil {
+			return types.ErrEmptyPublicKey(ModuleName)
 		}
 	}
+	// check for duplicate transaction to prevent replay attacks
+	txHash := tmTypes.Tx(txBz).Hash()
+	// make http call to tendermint to check txIndex // TODO remove this http call round trip and access directly
+	_, err := client.NewHTTP(tmNode.Config().RPC.ListenAddress, "/websocket").Tx(txHash, false)
+	if err == nil {
+		return types.ErrDuplicateTx(ModuleName, hex.EncodeToString(txHash))
+	}
+	// get the sign bytes from the tx
+	signBytes := GetSignBytes(ctx.ChainID(), stdTx)
+	// get the fees from the tx
+	expectedFee := sdk.NewCoins(sdk.NewCoin(sdk.DefaultStakeDenom, stdTx.GetMsg().GetFee()))
+	// test for public key type
+	p, ok := pk.(posCrypto.PublicKeyMultiSig)
+	// if standard public key
+	if !ok {
+		// validate the fees for a standard public key
+		if !stdTx.Fee.IsAllGTE(expectedFee) {
+			return types.ErrInsufficientFee(ModuleName, expectedFee, stdTx.Fee)
+		}
+		// validate signature for regular public key
+		if !simulate && !pk.VerifyBytes(signBytes, stdTx.Signature.Signature) {
+			return sdk.ErrUnauthorized("signature verification failed for the transaction")
+		}
+	}
+	// validate the signature depth
+	ok = ValidateSignatureDepth(params.TxSigLimit, p)
+	if !ok {
+		return types.ErrTooManySignatures(ModuleName, params.TxSigLimit)
+	}
+	// validate the multi sig
+	if !simulate && !pk.VerifyBytes(signBytes, stdTx.Signature.Signature) {
+		return sdk.ErrUnauthorized("multisignature verification failed for the transaction")
+	}
+	return nil
+}
 
-	return sdk.Result{}
+func ValidateSignatureDepth(limit uint64, publicKey posCrypto.PublicKeyMultiSig) (ok bool) {
+	_, ok = recSignDepth(1, limit, publicKey)
+	return
+}
+
+// recSignDepth ensures that the number of signatures does not exceed the max
+func recSignDepth(count, limit uint64, publicKey posCrypto.PublicKeyMultiSig) (c uint64, ok bool) {
+	for _, p := range publicKey.Keys() {
+		count++
+		if pk, ok := p.(posCrypto.PublicKeyMultiSig); ok {
+			count, ok = recSignDepth(count, limit, pk)
+			if !ok {
+				return count, ok
+			}
+		}
+		if count > limit {
+			return count, false
+		}
+	}
+	return count, true
+}
+
+// GetSignerAcc returns an account for a given address that is expected to sign
+// a transaction.
+func GetSignerAcc(ctx sdk.Ctx, ak keeper.Keeper, addr sdk.Address) (Account, sdk.Error) {
+	if acc := ak.GetAccount(ctx, addr); acc != nil {
+		return acc, nil
+	}
+	return nil, sdk.ErrUnknownAddress(fmt.Sprintf("account %s does not exist", addr))
 }
 
 // ValidateMemo validates the memo size.
-func ValidateMemo(stdTx StdTx, params Params) sdk.Result {
+func ValidateMemo(stdTx StdTx, params Params) sdk.Error {
 	memoLength := len(stdTx.GetMemo())
 	if uint64(memoLength) > params.MaxMemoCharacters {
 		return sdk.ErrMemoTooLarge(
@@ -138,112 +134,38 @@ func ValidateMemo(stdTx StdTx, params Params) sdk.Result {
 				"maximum number of characters is %d but received %d characters",
 				params.MaxMemoCharacters, memoLength,
 			),
-		).Result()
+		)
 	}
-
-	return sdk.Result{}
-}
-
-// verify the signature and increment the sequence. If the account doesn't have
-// a pubkey, set it.
-func processSig(acc Account, sig StdSignature, signBytes []byte, simulate bool) (updatedAcc Account, res sdk.Result) {
-	pubKey, res := ProcessPubKey(acc, sig)
-	if !res.IsOK() {
-		return nil, res
-	}
-	pk, ok := pubKey.(posCrypto.PublicKeyMultiSig)
-	if ok {
-		if !simulate && !pk.VerifyBytes(signBytes, sig.Signature) {
-			return nil, sdk.ErrUnauthorized("signature verification failed; verify correct account sequence and chain-id").Result()
-		}
-	}
-	if !simulate && !pubKey.VerifyBytes(signBytes, sig.Signature) {
-		return nil, sdk.ErrUnauthorized("signature verification failed; verify correct account sequence and chain-id").Result()
-	}
-	return acc, res
-}
-
-// ProcessPubKey verifies that the given account address matches that of the
-// StdSignature. In addition, it will set the public key of the account if it
-// has not been set.
-func ProcessPubKey(acc Account, sig StdSignature) (posCrypto.PublicKey, sdk.Result) {
-	// If pubkey is not known for account, set it from the StdSignature.
-	pubKey := acc.GetPubKey()
-	if pubKey == nil {
-		pubKey = sig.PublicKey
-		if pubKey == nil {
-			return nil, sdk.ErrInvalidPubKey("PubKey not found").Result()
-		}
-		if !bytes.Equal(pubKey.Address(), acc.GetAddress()) {
-			return nil, sdk.ErrInvalidPubKey(
-				fmt.Sprintf("PubKey does not match Signer address %s", acc.GetAddress())).Result()
-		}
-	}
-	return pubKey, sdk.Result{}
+	return nil
 }
 
 // DeductFees deducts fees from the given account.
-//
-// NOTE: We could use the CoinKeeper (in addition to the Keeper, because
-// the CoinKeeper doesn't give us accounts), but it seems easier to do this.
-func DeductFees(keeper keeper.Keeper, ctx sdk.Ctx, acc Account, fees sdk.Coins) sdk.Result {
-	blockTime := ctx.BlockHeader().Time
-	coins := acc.GetCoins()
-
+func DeductFees(keeper keeper.Keeper, ctx sdk.Ctx, tx types.StdTx) sdk.Error {
+	fees := tx.Fee
 	if !fees.IsValid() {
-		return sdk.ErrInsufficientFee(fmt.Sprintf("invalid fee amount: %s", fees)).Result()
+		return sdk.ErrInsufficientFee(fmt.Sprintf("invalid fee amount: %s", fees))
 	}
-
+	acc, err := GetSignerAcc(ctx, keeper, tx.GetSigner())
+	if err != nil {
+		return err
+	}
+	coins := acc.GetCoins()
 	// verify the account has enough funds to pay for fees
 	_, hasNeg := coins.SafeSub(fees)
 	if hasNeg {
-		return sdk.ErrInsufficientFunds(
-			fmt.Sprintf("insufficient funds to pay for fees; %s < %s", coins, fees),
-		).Result()
+		return types.ErrInsufficientBalance(ModuleName, acc.GetAddress(), fees)
 	}
-
-	// Validate the account has enough "spendable" coins as this will cover cases
-	// such as vesting accounts.
-	spendableCoins := acc.SpendableCoins(blockTime)
-	if _, hasNeg := spendableCoins.SafeSub(fees); hasNeg {
-		return sdk.ErrInsufficientFunds(
-			fmt.Sprintf("insufficient funds to pay for fees; %s < %s", spendableCoins, fees),
-		).Result()
-	}
-
-	err := keeper.SendCoinsFromAccountToModule(ctx, acc.GetAddress(), types.FeeCollectorName, fees)
+	err = keeper.SendCoinsFromAccountToModule(ctx, acc.GetAddress(), types.FeeCollectorName, fees)
 	if err != nil {
-		return err.Result()
+		return err
 	}
-
-	return sdk.Result{}
-}
-
-// EnsureSufficientMempoolFees verifies that the given transaction has supplied
-// enough fees to cover a proposer's minimum fees. A result object is returned
-// indicating success or failure.
-//
-// Contract: This should only be called during CheckTx as it cannot be part of
-// consensus.
-func EnsureSufficientMempoolFees(ctx sdk.Ctx, stdFee sdk.Coins) sdk.Result {
-	minGasPrices := ctx.MinGasPrices()
-	if !minGasPrices.IsZero() {
-		requiredFees := make(sdk.Coins, len(minGasPrices))
-		if !stdFee.IsAnyGTE(requiredFees) {
-			return sdk.ErrInsufficientFee(
-				fmt.Sprintf(
-					"insufficient fees; got: %q required: %q", stdFee, requiredFees,
-				),
-			).Result()
-		}
-	}
-	return sdk.Result{}
+	return nil
 }
 
 // GetSignBytes returns a slice of bytes to sign over for a given transaction
 // and an account.
 func GetSignBytes(chainID string, stdTx StdTx) []byte {
 	return StdSignBytes(
-		chainID, stdTx.Entropy, stdTx.Fee, stdTx.Msgs, stdTx.Memo,
+		chainID, stdTx.Entropy, stdTx.Fee, stdTx.Msg, stdTx.Memo,
 	)
 }